+++ yandex/resource_yandex_mdb_mysql_cluster.go
@@ -385,6 +385,13 @@ func resourceYandexMDBMySQLCluster() *schema.Resource {
 					},
 				},
 			},
+			"host_group_ids": {
+				Type:     schema.TypeSet,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
+				Optional: true,
+				Computed: true,
+			},
 		},
 	}
 }
@@ -478,6 +485,7 @@ func resourceYandexMDBMySQLClusterRestore(d *schema.ResourceData, meta interface
 		NetworkId:        req.NetworkId,
 		FolderId:         req.FolderId,
 		SecurityGroupIds: req.SecurityGroupIds,
+		HostGroupIds:     req.HostGroupIds,
 	}))
 	if err != nil {
 		return fmt.Errorf("Error while requesting API to create MySQL Cluster from backup %v: %s", backupID, err)
@@ -564,6 +572,7 @@ func prepareCreateMySQLRequest(d *schema.ResourceData, meta *Config) (*mysql.Cre
 	}
 
 	securityGroupIds := expandSecurityGroupIds(d.Get("security_group_ids"))
+	hostGroupIds := expandHostGroupIds(d.Get("host_group_ids"))
 
 	networkID, err := expandAndValidateNetworkId(d, meta)
 	if err != nil {
@@ -583,6 +592,7 @@ func prepareCreateMySQLRequest(d *schema.ResourceData, meta *Config) (*mysql.Cre
 		Labels:             labels,
 		SecurityGroupIds:   securityGroupIds,
 		DeletionProtection: d.Get("deletion_protection").(bool),
+		HostGroupIds:       hostGroupIds,
 	}
 	return &req, nil
 }
@@ -720,6 +730,10 @@ func resourceYandexMDBMySQLClusterRead(d *schema.ResourceData, meta interface{})
 	}
 
 	if err := d.Set("performance_diagnostics", perfDiag); err != nil {
+		return err
+	}
+
+	if err = d.Set("host_group_ids", cluster.HostGroupIds); err != nil {
 		return err
 	}
 
@@ -881,6 +895,9 @@ func getMysqlClusterUpdateRequest(d *schema.ResourceData) (*mysql.UpdateClusterR
 	}
 
 	securityGroupIds := expandSecurityGroupIds(d.Get("security_group_ids"))
+	if d.HasChange("host_group_ids") {
+		return nil, fmt.Errorf("host_group_ids change is not supported yet")
+	}
 
 	maintenanceWindow, err := expandMySQLMaintenanceWindow(d)
 	if err != nil {